few updates
